FIELD-2103: prevent catch_count = 0 and associated save errors
See https://www.fisheries.noaa.gov/jira/browse/FIELD-2103

Issues arise on the FG side when a PHLB record with biospecimens is activated in the CC entry page.  When adding a subsequent CC no PHLB, the Catches.catch_count value is set to 0, not None.  The 0 prevents the new Catch record to save properly, a constraint called CHK_CATCH_COUNT_VAL prevents 0 values.  Fix prevents this (doesn't fix the actual issue of sticky PHLB vals, but this was simple and non-invasive).